Made some refactoring.
