+++ src/test/java/suiteBuilder/functionalityContainer/LoginFunctionalityTest.java
@@ -1,8 +1,7 @@
-package suiteBuilder.abstractFunctionalityContainer;
+package suiteBuilder.functionalityContainer;
 
 import suiteBuilder.testAreaContainer.LoginTestArea;
 import suiteBuilder.testAreaContainer.TestArea;
-import suiteBuilder.testSuiteBuilder.TestAttributes;
 
 public abstract class LoginFunctionalityTest implements TestAttributes {
     @Override
