+++ src/test/java/suiteBuilder/functionalityContainer/SearchFunctionalityTest.java
@@ -1,8 +1,7 @@
-package suiteBuilder.abstractFunctionalityContainer;
+package suiteBuilder.functionalityContainer;
 
 import suiteBuilder.testAreaContainer.SearchTestArea;
 import suiteBuilder.testAreaContainer.TestArea;
-import suiteBuilder.testSuiteBuilder.TestAttributes;
 
 public abstract class SearchFunctionalityTest implements TestAttributes {
     @Override
