+++ src/test/java/suiteBuilder/testSuiteBuilder/TestSuiteTemplate.java
@@ -1,5 +1,7 @@
 package suiteBuilder.testSuiteBuilder;
 
+import suiteBuilder.functionalityContainer.TestAttributes;
+
 import java.util.ArrayList;
 import java.util.List;
